+++ src/betterdiscord/types/discord/modules.d.ts
@@ -84,6 +84,7 @@ export interface Dispatcher {
     dispatch<T>(payload: T): void;
     isDispatching(): boolean;
     subscribe<T extends (...args: any[]) => void>(id: string, cb: T): void;
+    addInterceptor<T>(callback: (payload: T) => void): string;
 }